fix(calculator): resolve database permissions and improve development workflow
🔧 Critical Fixes:
- Fixed calculator template 'not found' errors by granting proper database permissions
- Resolved 'permission denied for table' (PostgreSQL error 42501) on calculator tables
- Fixed invalid Supabase query syntax in client configurations and calculation history
- Updated config API to use Supabase directly for database consistency

📊 Calculator Improvements:
- Simplified database queries to use select('*') with post-processing
- Added proper field mapping (snake_case → camelCase) for API responses
- Improved error handling and logging for better debugging
- Ensured consistency across all calculator API endpoints

🗃️ Database:
- Added fix-calculator-permissions.sql migration
- Granted SELECT permissions to authenticated and anon roles
- Granted INSERT/UPDATE permissions for configurations and history

🧪 Development Experience:
- Added configurable logging via NEXT_PUBLIC_LOG_LEVEL and PRISMA_LOG_LEVEL
- Reduced console noise from Prisma queries in development
- Wrapped verbose logs in debug mode conditions for cleaner output
- Added environment-based log level control

📚 Documentation:
- Added CALCULATOR_TEMPLATE_FIX.md with root cause analysis
- Added DEVELOPMENT_LOGGING_CLEANUP.md for logging configuration
- Updated CALCULATOR_TESTS.md with comprehensive test cases
- Added SETUP_SUMMARY.md for Git workflow setup

🔄 Git Workflow Enhancements:
- Added PR template with tech stack checklist
- Created enhanced pre-commit and pre-push hooks
- Added setup scripts for automated Git hook installation
- Documented deployment plan and best practices

✅ Verification:
- All changes compile successfully (pnpm build passes)
- Calculator APIs now return proper responses
- Database permissions verified via Supabase MCP
- Logging configurable for development and production

Branch: hotfix/calculator-drives
Impact: Critical - Restores full calculator functionality
