+++ compare-hooks.sh
@@ -1,0 +1,83 @@
+#!/bin/bash
+
+# Quick comparison of your current vs enhanced hooks
+
+echo "📊 Hook Configuration Comparison"
+echo "=================================="
+echo ""
+
+echo "🔵 CURRENT PRE-COMMIT (.husky/pre-commit):"
+echo "  Status: DISABLED (exits 0 immediately)"
+echo "  Checks: None"
+echo ""
+
+echo "🟢 ENHANCED PRE-COMMIT (.husky/pre-commit.enhanced):"
+echo "  Status: ACTIVE"
+echo "  Checks:"
+echo "    ✅ TypeScript type checking (pnpm typecheck)"
+echo "    ✅ ESLint validation (pnpm lint)"
+echo "    ✅ Prisma schema validation (pnpm prisma validate)"
+echo "    ✅ Unit tests (pnpm test:unit --passWithNoTests)"
+echo "  Estimated time: 15-30 seconds"
+echo ""
+
+echo "🔵 CURRENT PRE-PUSH (.husky/pre-push):"
+echo "  Status: LENIENT (warnings only, doesn't fail)"
+echo "  Checks:"
+echo "    ⚠️  TypeScript check (informational only)"
+echo "    ⚠️  Build validation (commented out)"
+echo ""
+
+echo "🟢 ENHANCED PRE-PUSH (.husky/pre-push.enhanced):"
+echo "  Status: ACTIVE"
+echo "  Checks:"
+echo "    ✅ TypeScript type checking (pnpm typecheck)"
+echo "    ✅ ESLint validation (pnpm lint)"
+echo "    ✅ Prisma generate & validate"
+echo "    ✅ Unit tests (pnpm test:unit)"
+echo "    ✅ Integration tests (pnpm test:integration)"
+echo "    ✅ Production build (pnpm build:no-typecheck)"
+echo "    💡 E2E tests (optional, currently commented)"
+echo "  Estimated time: 1-3 minutes"
+echo ""
+
+echo "📋 NEW: PR TEMPLATE (.github/PULL_REQUEST_TEMPLATE.md):"
+echo "  ✨ Comprehensive checklist including:"
+echo "    - Type checking, linting, tests"
+echo "    - Database migration verification"
+echo "    - Security & performance checks"
+echo "    - Documentation requirements"
+echo "    - Reviewer notes"
+echo ""
+
+echo "⚖️  RECOMMENDATION:"
+echo ""
+echo "  Current setup is very permissive - good for rapid development"
+echo "  but may let bugs slip through to production."
+echo ""
+echo "  Enhanced setup provides safety net while staying fast enough"
+echo "  for daily development workflow."
+echo ""
+echo "  🎯 Suggested approach:"
+echo "     1. Start with enhanced pre-commit (catches most issues early)"
+echo "     2. Enable enhanced pre-push for critical branches"
+echo "     3. Use PR template for all PRs to main"
+echo "     4. Run full deploy:pre-checks before production deploys"
+echo ""
+
+read -p "Would you like to see the actual hook files? (y/n) " -n 1 -r
+echo ""
+if [[ $REPLY =~ ^[Yy]$ ]]; then
+    echo ""
+    echo "=== CURRENT PRE-COMMIT ==="
+    cat .husky/pre-commit
+    echo ""
+    echo "=== ENHANCED PRE-COMMIT ==="
+    cat .husky/pre-commit.enhanced
+    echo ""
+    echo "=== CURRENT PRE-PUSH ==="
+    cat .husky/pre-push
+    echo ""
+    echo "=== ENHANCED PRE-PUSH ==="
+    cat .husky/pre-push.enhanced
+fi
