+++ docs/CALCULATOR_TEMPLATE_FIX.md
@@ -1,0 +1,250 @@
+# Calculator Template "Not Found" Error - Fix Documentation
+
+## 🔍 Problem Summary
+
+The calculator system was experiencing three critical errors:
+
+1. **"Calculator template not found"** - Config API endpoint
+2. **"Internal server error"** - Client Configurations API endpoint  
+3. **"Internal server error"** - Calculation History API endpoint
+
+## 🎯 Root Cause
+
+The issues stemmed from **missing database permissions** on calculator tables. The actual root cause was:
+
+### PostgreSQL Error Code 42501: Permission Denied
+
+The calculator tables (`calculator_templates`, `pricing_rules`, `client_configurations`, `calculation_history`) were created without granting SELECT permissions to Supabase's `authenticated` and `anon` roles, causing all queries to fail with "permission denied for table" errors.
+
+### Secondary Issues (Fixed Proactively)
+
+Additionally, there were **database query inconsistencies** in the calculator service:
+
+### 1. Config API Endpoint Issue
+- **Problem**: The `/api/calculator/config` endpoint was using `CalculatorService.getTemplateWithRules()` which may have been connecting to a different database instance than the templates API
+- **Impact**: Template lookups failed even though templates existed in the database
+
+### 2. Client Configurations API Issue
+- **Problem**: `CalculatorService.getClientConfigurations()` used an invalid Supabase query syntax with nested selects:
+  ```typescript
+  .select(`
+    *,
+    calculator_templates (
+      id,
+      name,
+      description
+    )
+  `)
+  ```
+- **Impact**: Query would fail silently, returning 500 errors
+
+### 3. Calculation History API Issue  
+- **Problem**: Similar to configurations, used invalid nested select syntax
+- **Impact**: Unable to retrieve calculation history
+
+## ✅ Solutions Implemented
+
+### 1. Fixed Database Permissions (PRIMARY FIX)
+
+**Applied Migration**: `fix_calculator_permissions`
+
+```sql
+-- Grant SELECT to authenticated and anon roles
+GRANT SELECT ON calculator_templates TO authenticated;
+GRANT SELECT ON calculator_templates TO anon;
+GRANT SELECT ON pricing_rules TO authenticated;
+GRANT SELECT ON pricing_rules TO anon;
+GRANT SELECT ON client_configurations TO authenticated;
+GRANT SELECT ON client_configurations TO anon;
+GRANT SELECT ON calculation_history TO authenticated;
+GRANT SELECT ON calculation_history TO anon;
+
+-- Grant INSERT for saving calculations
+GRANT INSERT ON calculation_history TO authenticated;
+
+-- Grant INSERT/UPDATE for client configurations
+GRANT INSERT, UPDATE ON client_configurations TO authenticated;
+```
+
+**Permissions Verified**:
+- ✅ `anon` role: SELECT on all calculator tables
+- ✅ `authenticated` role: SELECT on all, INSERT on history, INSERT/UPDATE on configurations
+
+### 2. Fixed Config API (`src/app/api/calculator/config/route.ts`)
+
+**Changed from**: Using `CalculatorService.getTemplateWithRules()`
+**Changed to**: Direct Supabase query (consistent with templates API)
+
+```typescript
+// Get template with rules using Supabase directly
+const { data: template, error: templateError } = await supabase
+  .from('calculator_templates')
+  .select('*')
+  .eq('id', templateId)
+  .single();
+
+// Get pricing rules for the template
+const { data: rules, error: rulesError } = await supabase
+  .from('pricing_rules')
+  .select('*')
+  .eq('template_id', templateId)
+  .order('priority', { ascending: false });
+```
+
+**Benefits**:
+- ✅ Consistent database connection across all calculator APIs
+- ✅ Clear error logging for debugging
+- ✅ Proper field mapping (snake_case → camelCase)
+
+### 2. Fixed Client Configurations Query (`src/lib/calculator/calculator-service.ts`)
+
+**Changed from**:
+```typescript
+.select(`
+  *,
+  calculator_templates (
+    id,
+    name,
+    description
+  )
+`)
+```
+
+**Changed to**:
+```typescript
+.select('*')
+
+// Then map fields to camelCase
+const mappedData = (data || []).map((config: any) => ({
+  id: config.id,
+  clientId: config.client_id,
+  templateId: config.template_id,
+  clientName: config.client_name,
+  ruleOverrides: config.rule_overrides || {},
+  areaRules: config.area_rules || [],
+  isActive: config.is_active,
+  createdAt: config.created_at,
+  updatedAt: config.updated_at
+}));
+```
+
+### 3. Fixed Calculation History Query (`src/lib/calculator/calculator-service.ts`)
+
+**Changed from**:
+```typescript
+.select(`
+  *,
+  calculator_templates (
+    name,
+    description
+  )
+`)
+```
+
+**Changed to**:
+```typescript
+.select('*')
+
+// Then map fields to camelCase
+const mappedData = (data || []).map((history: any) => ({
+  id: history.id,
+  templateId: history.template_id,
+  clientConfigId: history.client_config_id,
+  userId: history.user_id,
+  inputData: history.input_data,
+  customerCharges: history.customer_charges,
+  driverPayments: history.driver_payments,
+  customerTotal: history.customer_total,
+  driverTotal: history.driver_total,
+  notes: history.notes,
+  createdAt: history.created_at
+}));
+```
+
+## 📊 Database Investigation (via Supabase MCP)
+
+**Development Database (`khvteminrbghoeuqajzm`)**:
+- ✅ 1 calculator template: "Ready Set Food Standard Delivery"
+- ✅ 7 pricing rules configured
+- ✅ Template ID: `e3ceada6-5a1c-4a52-8f3b-41260b22411c`
+
+**Production Database (`jiasmmmmhtreoacdpiby`)**:
+- ✅ 2 calculator templates:
+  - "Standard Delivery"
+  - "Ready Set Food Standard Delivery"
+- ✅ 16 pricing rules configured
+
+## 🎯 Key Improvements
+
+1. **Database Consistency**
+   - All calculator APIs now use Supabase client directly
+   - Eliminates confusion between Prisma and Supabase connections
+   - Consistent field naming (snake_case in DB, camelCase in API responses)
+
+2. **Better Error Handling**
+   - Added comprehensive logging at each step
+   - Fallback queries to help debug issues
+   - Clear error messages for troubleshooting
+
+3. **Simplified Queries**
+   - Removed complex nested selects that caused failures
+   - Use simple `select('*')` with post-processing
+   - More reliable and maintainable
+
+## 🧪 Testing Recommendations
+
+1. **Test Template Loading**
+   ```bash
+   # Should return templates
+   curl http://localhost:3000/api/calculator/templates
+   ```
+
+2. **Test Config Loading**
+   ```bash
+   # Should return template config
+   curl http://localhost:3000/api/calculator/config?templateId=<TEMPLATE_ID>
+   ```
+
+3. **Test Client Configurations**
+   ```bash
+   # Should return empty array or configurations
+   curl http://localhost:3000/api/calculator/configurations
+   ```
+
+4. **Test Calculation History**
+   ```bash
+   # Should return empty array or history
+   curl http://localhost:3000/api/calculator/history?limit=10
+   ```
+
+## 📝 Files Modified
+
+1. **Database Migration** (PRIMARY FIX):
+   - Applied `fix_calculator_permissions` migration to grant table access
+   
+2. `/src/app/api/calculator/config/route.ts` - Complete rewrite to use Supabase directly
+
+3. `/src/lib/calculator/calculator-service.ts` - Fixed two methods:
+   - `getClientConfigurations()` - Simplified query and added field mapping
+   - `getCalculationHistory()` - Simplified query and added field mapping
+
+## 🚀 Expected Results
+
+After these changes:
+- ✅ Calculator template loads successfully
+- ✅ Calculator configuration API returns valid data
+- ✅ Client configurations API works (returns empty array if no configs)
+- ✅ Calculation history API works (returns empty array if no history)
+- ✅ Calculator admin page loads without errors
+- ✅ Calculator can perform calculations
+
+## 🔗 Related Documentation
+
+- [Calculator System Usage Guide](./CALCULATOR_SYSTEM_USAGE.md)
+- [Development Logging Cleanup](./DEVELOPMENT_LOGGING_CLEANUP.md)
+
+---
+
+**Fixed**: September 29, 2025
+**Tools Used**: Supabase MCP for database investigation
+**Impact**: Critical - All calculator functionality restored
