+++ docs/DEVELOPMENT_LOGGING_CLEANUP.md
@@ -1,0 +1,92 @@
+# Development Logging Cleanup
+
+This document outlines the changes made to clean up development logs and reduce console noise during development.
+
+## Changes Made
+
+### 1. Prisma Logging Configuration
+
+**Problem**: Prisma was logging every SQL query (`prisma:query`), transaction start (`BEGIN`), and end (`COMMIT`) in development, flooding the console.
+
+**Solution**: Modified Prisma client instantiation to use environment variables for log level control:
+
+- **Files Modified**: 
+  - `src/lib/db/prisma-pooled.ts`
+  - `src/utils/prismaDB.ts`
+
+**New Behavior**:
+- **Development (default)**: Only logs warnings and errors (`['warn', 'error']`)
+- **Debug mode**: All logs when `NEXT_PUBLIC_LOG_LEVEL=debug` or `PRISMA_LOG_LEVEL` is set
+- **Production**: Only errors and warnings (`['error', 'warn']`)
+
+### 2. Custom Application Logs
+
+**Problem**: Verbose custom `console.log` statements throughout the application were creating noise during normal development.
+
+**Files Modified**:
+- `src/contexts/UserContext.tsx` - Wrapped verbose authentication and hydration logs
+- `src/app/actions/login.ts` - Wrapped detailed login flow logs
+- `src/app/api/users/current-user/route.ts` - Wrapped user profile dumps
+
+**Solution**: Wrapped verbose logs in environment variable conditions:
+```typescript
+if (process.env.NEXT_PUBLIC_LOG_LEVEL === 'debug') {
+  console.log('Detailed debug information...');
+}
+```
+
+### 3. Prisma Client Singleton Pattern
+
+**Status**: ✅ Already properly implemented
+- Main application uses singleton pattern via `src/lib/db/prisma-pooled.ts`
+- Debug routes intentionally create fresh clients for testing purposes (this is correct)
+
+## Environment Variables for Log Control
+
+### For Quiet Development (Recommended)
+Create a `.env.local` file with:
+```bash
+# Leave these unset or empty for quiet logs
+# NEXT_PUBLIC_LOG_LEVEL=
+# PRISMA_LOG_LEVEL=
+```
+
+### For Debugging Issues
+```bash
+# Enable verbose logging
+NEXT_PUBLIC_LOG_LEVEL=debug
+PRISMA_LOG_LEVEL=query,info,warn,error
+```
+
+### For Production-like Logging in Development
+```bash
+# Minimal logging
+NEXT_PUBLIC_LOG_LEVEL=
+PRISMA_LOG_LEVEL=warn,error
+```
+
+## Log Level Options
+
+### NEXT_PUBLIC_LOG_LEVEL
+- **Empty/unset**: Quiet development (recommended)
+- **`debug`**: Enable all custom application logs
+
+### PRISMA_LOG_LEVEL
+- **Empty/unset**: Use defaults (warn,error in dev; error,warn in prod)
+- **Custom**: Comma-separated list from: `query`, `info`, `warn`, `error`
+
+## Benefits
+
+1. **Cleaner Development Console**: Reduced noise from database queries and verbose debug logs
+2. **Flexible Debugging**: Easy to enable detailed logs when investigating issues
+3. **Better Performance**: Fewer console operations during normal development
+4. **Production Ready**: Minimal logging overhead in production environments
+
+## Usage
+
+- **Normal Development**: No environment variables needed (quiet by default)
+- **Debugging Authentication**: Set `NEXT_PUBLIC_LOG_LEVEL=debug`
+- **Debugging Database**: Set `PRISMA_LOG_LEVEL=query,info,warn,error`
+- **Full Debug Mode**: Set both variables for maximum verbosity
+
+The logging system now provides a clean development experience while maintaining the ability to debug when needed.
